+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - All `id` fields default to `bigint` thanks to `rails-bigint-pk` gem
+- Move `Rails.configuration.x.shopify_app_name` to `DiscoApp.configuration.app_name`
 
 ## 0.8.3 - 2016-02-04
 ### Changed